Added Mezzio development config discovery and injection
